Changes buffers to float

+++ shaders/average.glsl
@@ -2,16 +2,16 @@
 
 layout (local_size_x = 8, local_size_y = 4, local_size_z = 1) in;
 layout (binding = EFTDEM_HEIGHTMAP_BUFFER) restrict buffer heightsBuffer{
-    double heights[];
+    float heights[];
 };
 layout (binding = EFTDEM_SUM_BUFFER) restrict buffer sumsBuffer{
-    double sums[];
+    float sums[];
 };
 layout (binding = EFTDEM_AMOUNT_BUFFER) restrict buffer amountsBuffer{
     uint amounts[];
 };
 layout (binding = EFTDEM_AVERAGE_BUFFER) restrict buffer averagesBuffer{
-    double average[];
+    float average[];
 };
 
 uniform uvec2 resolution;
@@ -28,5 +28,9 @@ void main() {
     uint coord1D;
 
     coord1D = calculate1DCoordinate(correctedGlobalInvocation);
-    average[coord1D] = (heights[coord1D] > 0) ? heights[coord1D] : (amounts[coord1D] == 0) ? 0.0 : sums[coord1D]/amounts[coord1D];
+    average[coord1D] = (heights[coord1D] > 0.)
+        ? heights[coord1D]
+        : ((amounts[coord1D] == 0)
+            ? 0.0
+            : sums[coord1D] / amounts[coord1D]);
 }
