+++ shaders/horizontalSum.glsl
@@ -2,10 +2,10 @@
 
 layout (local_size_x = 8, local_size_y = 4, local_size_z = 1) in;
 layout (binding = EFTDEM_HEIGHTMAP_BUFFER) restrict buffer mapBuffer{
-    double discreteValues[];
+    float discreteValues[];
 };
 layout (binding = EFTDEM_HORIZONTAL_SUM_BUFFER) restrict buffer horizontalSumBuffer{
-    double horizontalSum[];
+    float horizontalSum[];
 };
 
 uniform uvec2 resolution;
@@ -21,9 +21,9 @@ void main() {
     if (any(greaterThanEqual(correctedGlobalInvocation, resolution))) return;
     uint coord1D = calculate1DCoordinate(correctedGlobalInvocation);
 
-    double sum = 0.0;
+    float sum = 0.0;
 
-    for (uint kx = 0; kx <= 2*kernelRadius; kx++) {
+    for (uint kx = 0; kx <= 2 * kernelRadius; kx++) {
         uint x = kx - kernelRadius + correctedGlobalInvocation.x;
         uint y = correctedGlobalInvocation.y;
